rename: 파일명 변경 파일명, 컴포넌트 직관적으로 변경했습니다.
